refactor: decompose mission-executor (gating + generic collection mutator)
